backend/requirements.txt
Django>=4.2.0,<5.0.0
django-cors-headers>=4.3.0
django-rest-framework>=3.14.0
psycopg2-binary>=2.9.9
Telethon>=1.33.1
celery>=5.3.6
redis>=5.0.1
Python-dotenv>=1.0.0
Google-generativeai>=0.3.2
pydantic>=2.6.1
django-celery-beat>=2.5.0
django-celery-results>=2.5.1
flower>=2.0.1